requirements.txt
python-telegram-bot==13.15
flask==2.3.2
requests==2.31.0